[fix] Fix RewardModelTrainer API and tensor indexing
Fixes for CAI integration test failures:

1. RewardModelTrainer.train() parameter name
   - Change from 'preference_data' to 'training_data' in pipeline.py and tests
   - Matches actual API signature in reward_model.py:540

2. RewardModel tensor indexing bug (reward_model.py:101)
   - Cast sequence_lengths to .long() before using as tensor index
   - Fixes: IndexError: tensors used as indices must be long, int, byte or bool tensors

These fixes address the TypeError and IndexError in CAI integration tests.

+++ src/safety/constitutional/reward_model.py
@@ -98,7 +98,7 @@ class RewardModel(nn.Module):
 
         # Get hidden state of last token for each sequence in batch
         # We find the last non-padding token for each sequence
-        sequence_lengths = attention_mask.sum(dim=1) - 1  # -1 for 0-indexing
+        sequence_lengths = (attention_mask.sum(dim=1) - 1).long()  # -1 for 0-indexing, cast to long for indexing
         batch_size = hidden_states.shape[0]
 
         # Index to get last token hidden state for each sequence
